Allow field names to include a single `::` between two idents.
This will be used for a dynamic map type that uses fields as the keys
where some keys are required and some of those keys are namespaced at a
single level. We could model this as a map but modeling this as a struct
type in bauble allows easily specifing a limited set of key values that
are allowed with an optional set of required keys.